refactor metrics processing
