+++ main.go
@@ -1,17 +1,12 @@
 package main
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"os"
-	"sort"
 	"strings"
-	"time"
 
 	"github.com/Knetic/govaluate"
-	dto "github.com/prometheus/client_model/go"
-	"github.com/prometheus/common/expfmt"
 	"github.com/sensu-community/sensu-plugin-sdk/sensu"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/shirou/gopsutil/process"
@@ -195,333 +190,11 @@ func executeCheck(event *corev2.Event) (int, error) {
 		fmt.Printf("# Status - %s\n", mapSeverity(overallSeverity))
 	}
 	// Construct metrics
-	nowMS := time.Now().UnixMilli()
-	families := make([]*dto.MetricFamily, 0)
-	var processesFamily *dto.MetricFamily
-	processesFamily = newMetricFamily("processes", "SumoLogic Dashboard Compatible Cumulative Process Metrics", dto.MetricType_GAUGE)
-	families = append(families, processesFamily)
-	var procstatFamily *dto.MetricFamily
-	procstatFamily = newMetricFamily("procstat", "SumoLogic Dashboard Compatible Per Process Metrics", dto.MetricType_GAUGE)
-	families = append(families, procstatFamily)
-	totalProcesses := int64(0)
-	totalThreads := int64(0)
-	totalStatusMap := map[string]int64{
-		"parked":     0,
-		"wait":       0,
-		"blocked":    0,
-		"zombies":    0,
-		"dead":       0,
-		"stopped":    0,
-		"running":    0,
-		"sleeping":   0,
-		"idle":       0,
-		"unknown":    0,
-		"unexpected": 0,
+	err = generateMetrics(found)
+	if err != nil {
+		return sensu.CheckStateCritical, nil
 	}
-	totalTags := make(map[string]string)
-	hostname, err := os.Hostname()
-	if err == nil {
-		totalTags["host.name"] = hostname
-	}
-	for searchStr, processes := range found {
-		totalProcesses += int64(len(processes))
-		for _, p := range processes {
-			var fval32 float32
-			var fval64 float64
-			var ival32 int32
-			var ival64 int64
-
-			name, err := p.Name()
-			if err != nil {
-				continue
-			}
-			status, err := p.Status()
-			switch status[0] {
-			case 'P':
-				totalStatusMap["parked"] += 1
-			case 'W':
-				totalStatusMap["wait"] += 1
-			case 'U', 'D', 'L':
-				totalStatusMap["blocked"] += 1
-				// Also known as uninterruptible sleep or disk sleep
-			case 'Z':
-				totalStatusMap["zombies"] += 1
-			case 'X':
-				totalStatusMap["dead"] += 1
-			case 'T':
-				totalStatusMap["stopped"] += 1
-			case 'R':
-				totalStatusMap["running"] += 1
-			case 'S':
-				totalStatusMap["sleeping"] += 1
-			case 'I':
-				totalStatusMap["idle"] += 1
-			case '?':
-				totalStatusMap["unknown"] += 1
-			default:
-				totalStatusMap["unexpected"] += 1
-			}
-			procstatTags := make(map[string]string)
-			//Set the labels for the metric
-			hostname, err := os.Hostname()
-			if err == nil {
-				procstatTags["host.name"] = hostname
-			}
-			procstatTags["field"] = "none"
-			procstatTags["search_string"] = searchStr
-			procstatTags["process.executable.name"] = name
-
-			//cpu_usage metric
-			procstatTags["field"] = "cpu_usage"
-			procstatTags["units"] = "percent"
-			fval64, _ = p.CPUPercent()
-			gauge := newGaugeMetric(procstatTags, float64(fval64), nowMS)
-			procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			//memory_usage metric
-			procstatTags["field"] = "memory_usage"
-			procstatTags["units"] = "percent"
-			fval32, _ = p.MemoryPercent()
-			gauge = newGaugeMetric(procstatTags, float64(fval32), nowMS)
-			procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			mem, err := p.MemoryInfo()
-			if err == nil {
-				procstatTags["units"] = "bytes"
-				procstatTags["field"] = "memory_rss"
-				gauge = newGaugeMetric(procstatTags, float64(mem.RSS), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				procstatTags["field"] = "memory_vms"
-				gauge = newGaugeMetric(procstatTags, float64(mem.VMS), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				procstatTags["field"] = "memory_swap"
-				gauge = newGaugeMetric(procstatTags, float64(mem.Swap), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				procstatTags["field"] = "memory_data"
-				gauge = newGaugeMetric(procstatTags, float64(mem.Data), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				procstatTags["field"] = "memory_stack"
-				gauge = newGaugeMetric(procstatTags, float64(mem.Stack), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				procstatTags["field"] = "memory_locked"
-				gauge = newGaugeMetric(procstatTags, float64(mem.Locked), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			}
-			//created_at metric
-			procstatTags["field"] = "created_at"
-			procstatTags["units"] = "nanoseconds"
-			ival64, _ = p.CreateTime()
-			gauge = newGaugeMetric(procstatTags, float64(ival64)*1e6, nowMS)
-			procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			//num_fds metric
-			procstatTags["field"] = "num_fds"
-			procstatTags["units"] = "count"
-			ival32, _ = p.NumFDs()
-			gauge = newGaugeMetric(procstatTags, float64(ival32), nowMS)
-			procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			//num_threads metric
-			procstatTags["field"] = "num_threads"
-			procstatTags["units"] = "count"
-			ival32, _ = p.NumThreads()
-			totalThreads += int64(ival32)
-			gauge = newGaugeMetric(procstatTags, float64(ival32), nowMS)
-			procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			iostats, err := p.IOCounters()
-			if err == nil && iostats != nil {
-				//read_count metric
-				procstatTags["field"] = "read_count"
-				procstatTags["units"] = "count"
-				gauge = newGaugeMetric(procstatTags, float64(iostats.ReadCount), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				//read_bytes metric
-				procstatTags["field"] = "read_bytes"
-				procstatTags["units"] = "bytes"
-				gauge = newGaugeMetric(procstatTags, float64(iostats.ReadBytes), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				//write_count metric
-				procstatTags["field"] = "write_count"
-				procstatTags["units"] = "count"
-				gauge = newGaugeMetric(procstatTags, float64(iostats.WriteCount), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				//write_bytes metric
-				procstatTags["field"] = "write_bytes"
-				procstatTags["units"] = "bytes"
-				gauge = newGaugeMetric(procstatTags, float64(iostats.WriteBytes), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			}
-			faults, err := p.PageFaults()
-			if err == nil && faults != nil {
-				//major_fault metric
-				procstatTags["field"] = "major_faults"
-				procstatTags["units"] = "count"
-				gauge = newGaugeMetric(procstatTags, float64(faults.MajorFaults), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				//minor_fault metric
-				procstatTags["field"] = "minor_faults"
-				procstatTags["units"] = "count"
-				gauge = newGaugeMetric(procstatTags, float64(faults.MinorFaults), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				//child_major_fault metric
-				procstatTags["field"] = "child_major_faults"
-				procstatTags["units"] = "count"
-				gauge = newGaugeMetric(procstatTags, float64(faults.ChildMajorFaults), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				//child_minor_fault metric
-				procstatTags["field"] = "child_minor_faults"
-				procstatTags["units"] = "count"
-				gauge = newGaugeMetric(procstatTags, float64(faults.ChildMinorFaults), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			}
-			switches, err := p.NumCtxSwitches()
-			if err == nil && switches != nil {
-				//involuntary_context_switches
-				procstatTags["field"] = "involuntary_context_switches"
-				procstatTags["units"] = "count"
-				gauge = newGaugeMetric(procstatTags, float64(switches.Involuntary), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-				//voluntary_context_switches
-				procstatTags["field"] = "voluntary_context_switches"
-				procstatTags["units"] = "count"
-				gauge = newGaugeMetric(procstatTags, float64(switches.Voluntary), nowMS)
-				procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-			}
-			rlims, err := p.RlimitUsage(true)
-			if err == nil {
-				for _, rlim := range rlims {
-					procstatTags["units"] = "N/A"
-					var name string
-					switch rlim.Resource {
-					case process.RLIMIT_CPU:
-						name = "cpu_time"
-					case process.RLIMIT_DATA:
-						name = "memory_data"
-						procstatTags["units"] = "bytes"
-					case process.RLIMIT_STACK:
-						name = "memory_stack"
-						procstatTags["units"] = "bytes"
-					case process.RLIMIT_RSS:
-						name = "memory_rss"
-						procstatTags["units"] = "bytes"
-					case process.RLIMIT_NOFILE:
-						name = "num_fds"
-						procstatTags["units"] = "count"
-					case process.RLIMIT_MEMLOCK:
-						name = "memory_locked"
-						procstatTags["units"] = "bytes"
-					case process.RLIMIT_AS:
-						name = "memory_vms"
-						procstatTags["units"] = "bytes"
-					case process.RLIMIT_LOCKS:
-						name = "file_locks"
-						procstatTags["units"] = "count"
-					case process.RLIMIT_SIGPENDING:
-						name = "signals_pending"
-						procstatTags["units"] = "count"
-					case process.RLIMIT_NICE:
-						name = "nice_priority"
-					case process.RLIMIT_RTPRIO:
-						name = "realtime_priority"
-					default:
-						continue
-					}
-					procstatTags["field"] = "rlimit_" + name + "_soft"
-					gauge = newGaugeMetric(procstatTags, float64(rlim.Soft), nowMS)
-					procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-					procstatTags["field"] = "rlimit_" + name + "_hard"
-					gauge = newGaugeMetric(procstatTags, float64(rlim.Hard), nowMS)
-					procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-
-					switch rlim.Resource {
-					case process.RLIMIT_CPU,
-						process.RLIMIT_LOCKS,
-						process.RLIMIT_SIGPENDING,
-						process.RLIMIT_NICE,
-						process.RLIMIT_RTPRIO:
-						procstatTags["field"] = name
-						gauge = newGaugeMetric(procstatTags, float64(rlim.Used), nowMS)
-						procstatFamily.Metric = append(procstatFamily.Metric, gauge)
-					}
-
-				}
-			}
-		}
-	}
-	totalTags["field"] = "total"
-	totalTags["units"] = "count"
-	gauge := newGaugeMetric(totalTags, float64(totalProcesses), nowMS)
-	processesFamily.Metric = append(processesFamily.Metric, gauge)
-	totalTags["field"] = "total_threads"
-	totalTags["units"] = "count"
-	gauge = newGaugeMetric(totalTags, float64(totalThreads), nowMS)
-	processesFamily.Metric = append(processesFamily.Metric, gauge)
-	for status, count := range totalStatusMap {
-		totalTags["field"] = status
-		totalTags["units"] = "count"
-		gauge = newGaugeMetric(totalTags, float64(count), nowMS)
-		processesFamily.Metric = append(processesFamily.Metric, gauge)
-	}
-
-	var buf bytes.Buffer
-	for _, family := range families {
-		buf.Reset()
-		encoder := expfmt.NewEncoder(&buf, expfmt.FmtText)
-		err = encoder.Encode(family)
-		if err != nil {
-			return sensu.CheckStateCritical, err
-		}
-
-		fmt.Print(buf.String())
-	}
-
 	return overallSeverity, nil
-}
-func newGaugeMetric(tags map[string]string, value float64, timestampMS int64) *dto.Metric {
-	labels := make([]*dto.LabelPair, 0)
-	keys := make([]string, 0, len(tags))
-	for k := range tags {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	for _, key := range keys {
-		name := fmt.Sprintf("%v", key)
-		val := fmt.Sprintf("%v", tags[key])
-		labels = append(labels, &dto.LabelPair{Name: &name, Value: &val})
-	}
-	gauge := &dto.Metric{
-		Label: labels,
-		Gauge: &dto.Gauge{
-			Value: &value,
-		},
-		TimestampMs: &timestampMS,
-	}
-	return gauge
-}
-func newCounterMetric(tags map[string]string, value float64, timestampMS int64) *dto.Metric {
-	labels := make([]*dto.LabelPair, 0)
-	keys := make([]string, 0, len(tags))
-	for k := range tags {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	for _, key := range keys {
-		name := fmt.Sprintf("%v", key)
-		val := fmt.Sprintf("%v", tags[key])
-		labels = append(labels, &dto.LabelPair{Name: &name, Value: &val})
-	}
-	counter := &dto.Metric{
-		Label: labels,
-		Counter: &dto.Counter{
-			Value: &value,
-		},
-		TimestampMs: &timestampMS,
-	}
-	return counter
-}
-func newMetricFamily(name, help string, metricType dto.MetricType) *dto.MetricFamily {
-	return &dto.MetricFamily{
-		Name:   &name,
-		Help:   &help,
-		Type:   &metricType,
-		Metric: []*dto.Metric{},
-	}
 }
 
 func parseSearches(searchJSON string) ([]Search, error) {
